+++ test/bshb-utils.spec.ts
@@ -1,15 +1,15 @@
-import { BshbUtils } from "../src";
-import { expect } from "chai";
+import { BshbUtils } from '../src';
+import { expect } from 'chai';
 
-describe("BshbUtils", () => {
-  it("should generate identifier", () => {
+describe('BshbUtils', () => {
+  it('should generate identifier', () => {
     const identifier = BshbUtils.generateIdentifier();
 
     expect(identifier).to.not.be.null;
     expect(identifier.length).to.be.greaterThan(0);
   });
 
-  it("should generate client certificate", () => {
+  it('should generate client certificate', () => {
     const result = BshbUtils.generateClientCertificate();
 
     expect(result).to.not.be.null;